Randomly display each question's answers
To accomplish this, answers for all states are shuffled when site is first loaded. Since this is only done at this time (these data objectsare defined in src\App.tsx, but outside of the component), and not every time src\Components\State\State.tsx is rendered, options are not randomized after user answers the question.